+++ examples/transactions.rs
@@ -1,4 +1,4 @@
-use tarantool_rs::{Connection, ConnectionLike, Executor, Value};
+use tarantool_rs::{Connection, Executor, ExecutorExt, Value};
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {