+++ src/utils/mod.rs
@@ -1,3 +1,5 @@
+//! Various helpers.
+
 pub use unique_id_name_map::UniqueIdNameMap;
 
 pub(crate) use unique_id_name_map::UniqueIdName;